Multiple changes.
 * GetFleetspeakPendingMessages: correct handling of fleetspeak returning an empty/missing client ID.
 * Removed pure Python varints parsing.
 * Show ArtifactCollectorFlowProgress and FlowState in new UI.
 * Remove FlowListEntry and FlowResultSet.
 * Decouple flow results from flow details base class.
 * Update NPM dependencies, notably Angular from 11 to 12.
 * Sandboxing: correctly handle short reads/writes when reading/writing from a pipe.
 * Circuit break Interrogate on ClientStartup if recent Interrogate is still running.
 * Timeline test: support /bin symlinked to /usr/bin
 * Add support for escaping non-printable chars in the body export in the new UI.
 * Add support for escaping non-printable chars for the body export in API client.
 * Add support for escaping non-printable characters in the body export API method.
 * Add support for escaping non-printable characters in the body format util.
 * Remove FlowResults from ClientPageGlobalStore, replace with FlowResultsLocalStore and LoadFlowResultsDirective.
 * Interrogate clients when their version changes.
 * Add support for carriage return escape for body export in the new UI.
 * Add support for carriage return escape for the body export in API client.
 * Add support for carriage return escape in the body export API method.
 * Add support for carriage return escape in export utility for the body format.
 * MSI installer: integrate into open source continuous integration.
 * MSI: multiple changes.
 * MSI Installer: restart fleetspeak only in fleetspeak-enabled configuration.
 * Fix Angular compilation warnings.

+++ appveyor/windows_templates/build_windows_templates.py
@@ -9,6 +9,7 @@ import glob
 import os
 import shutil
 import subprocess
+import time
 
 parser = argparse.ArgumentParser(description="Build windows templates.")
 
@@ -86,7 +87,17 @@ parser.add_argument(
     help="Optional path to a 32-bit Python virtualenv to be used instead "
     "of creating a new one.")
 
+parser.add_argument(
+    "--build_msi",
+    dest="build_msi",
+    default=False,
+    action="store_true",
+    help="Enable building of an MSI template.")
+
 args = parser.parse_args()
+
+
+_SC_STOP_WAIT_TIME_SECS = 5
 
 
 class WindowsTemplateBuilder(object):
@@ -240,9 +251,23 @@ class WindowsTemplateBuilder(object):
       ]
     else:
       build_args = ["--verbose", "build", "--output", args.output_dir]
+    if args.build_msi:
+      wix_tools_path = self._WixToolsPath()
+      build_args += [
+          "-p",
+          "ClientBuilder.wix_tools_path=%{" + wix_tools_path + "}",
+          "-p",
+          "ClientBuilder.build_msi=true",
+      ]
     subprocess.check_call([self.grr_client_build64] + build_args)
     if args.build_32:
       subprocess.check_call([self.grr_client_build32] + build_args)
+
+  def _WixToolsPath(self) -> str:
+    matches = glob.glob("C:\\Program Files*\\WiX Toolset*")
+    if not matches:
+      raise Exception("Couldn't find WiX Toolset.")
+    return matches[0]
 
   def _RepackTemplates(self):
     """Repack templates with a dummy config."""
@@ -280,13 +305,32 @@ class WindowsTemplateBuilder(object):
 
   def _CleanupInstall(self):
     """Cleanup from any previous installer enough for _CheckInstallSuccess."""
-    if os.path.exists(self.install_path):
-      shutil.rmtree(self.install_path)
-      if os.path.exists(self.install_path):
-        raise RuntimeError("Install path still exists: %s" % self.install_path)
 
-    # Deliberately don't check return code, since service may not be installed.
-    subprocess.call(["sc", "stop", self.service_name])
+    subprocess.check_call(["sc", "stop", self.service_name])
+
+    if args.build_msi:
+      msiexec_args = [
+          "msiexec",
+          "/q",
+          "/x",
+          glob.glob(os.path.join(args.output_dir, "dbg_*_amd64.msi")).pop(),
+      ]
+      print("Running:", msiexec_args)
+      subprocess.check_call(msiexec_args)
+    else:
+      # Wait for service to stop.
+      for _ in range(_SC_STOP_WAIT_TIME_SECS):
+        output = subprocess.check_output(["sc", "query", self.service_name],
+                                         encoding="utf-8")
+        stopped = "STOPPED" in output
+        if stopped:
+          break
+        time.sleep(1.0)
+      if os.path.exists(self.install_path):
+        shutil.rmtree(self.install_path)
+        if os.path.exists(self.install_path):
+          raise RuntimeError("Install path still exists: %s" %
+                             self.install_path)
 
   def _CheckInstallSuccess(self):
     """Checks if the installer installed correctly."""
@@ -320,12 +364,21 @@ class WindowsTemplateBuilder(object):
     """Install the installer built by RepackTemplates."""
     # 32 bit binary will refuse to install on a 64bit system so we only install
     # the 64 bit version
-    installer_amd64 = glob.glob(
-        os.path.join(args.output_dir, "dbg_*_amd64.exe")).pop()
-    self._CleanupInstall()
+    if args.build_msi:
+      installer_amd64_args = [
+          "msiexec",
+          "/i",
+          glob.glob(os.path.join(args.output_dir, "dbg_*_amd64.msi")).pop(),
+      ]
+    else:
+      installer_amd64_args = [
+          glob.glob(os.path.join(args.output_dir, "dbg_*_amd64.exe")).pop()
+      ]
     # The exit code is always 0, test to see if install was actually successful.
-    subprocess.check_call([installer_amd64])
+    print("Running:", installer_amd64_args)
+    subprocess.check_call(installer_amd64_args)
     self._CheckInstallSuccess()
+    self._CleanupInstall()
 
   def Build(self):
     """Build templates."""
